requirements.txt
requests==2.32.3
web3==7.2.0
python-dotenv==1.0.1
Flask==3.0.3
pandas==2.2.2
pytest==7.4.0
